# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD Client API",
        "description": "Documentação da API de clientes via Swagger",
        "contact": {
            "email": "phmsanttos@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/customers": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar clientes",
                "operationId": "1b873d7b0c63876f30a978b52c4d24d2",
                "responses": {
                    "200": {
                        "description": "Lista de clientes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Criar cliente",
                "operationId": "747630c757eae4c38404384330e68c57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fulano da Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "fulano@email.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "(81) 99999-9999"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Recife"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "PE"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "https://example.com/foto.jpg"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso"
                    }
                }
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Mostrar cliente específico",
                "operationId": "d03d1278b7e26322e9a425406bc4e9f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualizar cliente",
                "operationId": "32da1527afa82d1a6b0e034146f1032d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Novo Nome"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "novo@email.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "(81) 98888-8888"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Olinda"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "PE"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "https://example.com/photo.jpg"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 25
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Deletar cliente",
                "operationId": "e9f16ac94d63309225711a1f3dfb245a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cliente deletado com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Clientes",
            "description": "Clientes"
        }
    ]
}